pyproject.toml
[project]
name = "0411-project-interview-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "braintree>=4.35.0",
    "django>=5.2",
    "django-environ>=0.12.0",
    "psycopg[binary]>=3.2.7",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
]
